requirements.txt
beautifulsoup4==4.13.1
boto3==1.36.11
botocore==1.36.11
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
jmespath==1.0.1
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
sib-api-v3-sdk==7.6.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
